backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "boto3>=1.42.10",
    "cryptography>=46.0.3",
    "email-validator>=2.3.0",
    "fastapi>=0.124.4",
    "google-generativeai>=0.8.5",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "inngest>=0.5.13",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "pgvector>=0.3.6",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "sse-starlette>=3.0.3",
    "uvicorn>=0.38.0",
    "vecs>=0.4.5",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.9",
]
